Casual drivers: split loads at 0.5, mine bonus from Group A, CTC incl. food
Three fixes to the casual driver pay cycle, all matching the wage sheet the
module replicates (4 split lines at R1100 = "4 LOADS", bonus 600 = 4 x 150,
cost to company = loads + bonus + food while wages = loads + bonus - food).

Split loads shown as whole loads:
- Drivers list "Loads (period)" counted only non-split loads, so a casual with
  1 full + 1 split showed 1 instead of 1.5. Added a batched _month_split_counts
  keyed by (driver_id, entity_id) and bucketed by statement period, matching the
  pay-cycle sync; DriverSummary.load_count_this_month is now a float.
- Payslip Summary printed the full-load count against an amount that already
  carried the split halves - now effective counts (effA/effB).
- Load Entry split notes spell out "2 x 1/2 = 1.0".

Mine bonus:
- Casual bonus now derives from Group A (group_a + 0.5 x split_group_a) instead
  of assmang_loads. Group A IS the bonus-mine set, and assmang_loads is only ever
  written by the TruckLoad syncs - hand-typed Group A loads earned no bonus, and
  stale values paid a bonus on cycles with no loads. Net +R5400 over 15 casual
  cycles (Julian Olifant May +1200, Johnny Jun +1050; HERMANUS BKMO May -450 had
  no loads at all). Permanent drivers keep the columns - they have no group buckets.
- The bonus is now a visible, overridable row in the Payslip Summary for casuals.

CTC:
- Casual CTC = gross + food allowance. Food stays a deduction from wages, so net
  payable is unchanged; it is NOT an earnings line (that would lift take-home pay).

Verified by replaying all 263 pay cycles through the old and new calculator:
0 permanent cycles changed.

+++ backend/app/schemas/schemas.py
@@ -1023,7 +1023,8 @@ class DriverSummary(BaseModel):
     truck_registration: Optional[str] = None
     subcontractor_name: Optional[str] = None
     is_active: bool
-    load_count_this_month: int = 0
+    # Float, not int — split loads count 0.5 each, so a driver can be on 3.5 loads.
+    load_count_this_month: float = 0
     # Both come from the driver's DriverPayCycle for the requested period
     # (the list's month/year params, defaulting to the current month).
     net_pay_this_month: Decimal = Decimal("0")
